🔧 refactor(mobile/utils): move content formatters to utils for better organization and reusability
