+++ apps/mobile-queue/utils/content-formatters.ts
